requirements.txt
PyAutoGUI
SendKeys
Pillow
python-dotenv
pywin32